requirements.txt
tensorflow==2.12.0
numpy
pandas
scikit-learn
pillow
streamlit
fpdf
